requirements.txt
openai
langchain
PyPDF2
streamlit
python-dotenv

